requirements.txt
numpy==1.19.0
opencv-python
matplotlib
scipy
pyefd
PyYAML
